move validate_phone_datetime to casexml.case
