+++ corehq/ex-submodules/casexml/apps/case/tests/test_strict_datetimes.py
@@ -1,10 +1,11 @@
 from xml.etree import ElementTree
 from django.test import TestCase
+
+from casexml.apps.case.exceptions import PhoneDateValueError
 from casexml.apps.case.util import make_form_from_case_blocks
 from casexml.apps.case.xform import process_cases_with_casedb
 from corehq.form_processor.interfaces.processor import FormProcessorInterface
 from corehq.form_processor.test_utils import run_with_all_backends
-from couchforms.exceptions import PhoneDateValueError
 from couchforms.util import process_xform
